requirements.txt
streamlit
gspread
oauth2client
requests
pandas
python-dotenv
openpyxl
